+++ tck/src/main/java/ee/jakarta/tck/data/framework/read/only/CustomRepository.java
@@ -34,7 +34,7 @@ public interface CustomRepository {
     @Insert
     void add(List<NaturalNumber> list);
 
-    int countByIdIn(Set<Long> ids);
+    long countByIdIn(Set<Long> ids);
 
     boolean existsByIdIn(Set<Long> ids);
